feat: power special method in modulo 12 class
Reflected power not implemented, because powering an integer or modulo 12 integer to a modulo 12 integer is not well-defined.

We have not used `pow(base, exp, mod=12)` in the method definition, because it forces integer base and exponent (see <https://docs.python.org/3/library/functions.html#pow>), and for now we allow floating point representatives.

Also tests.

Issue #1